# docs/swagger.json
{
    "components": {
        "schemas": {
            "activity.FeedResponse": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/domain.Activity"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/pagination.Meta"
                    }
                },
                "type": "object"
            },
...
            "domain.Group": {
                "properties": {
                    "archivedAt": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "createdBy": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "inviteCode": {
                        "type": "string"
                    },
                    "inviteCodeExpiresAt": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "requireAdminApproval": {
                        "type": "boolean"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "domain.JoinResponse": {
                "properties": {
                    "group": {
                        "$ref": "#/components/schemas/domain.Group"
                    },
                    "message": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "domain.Preview": {
                "properties": {
                    "creatorName": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "memberCount": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "requireAdminApproval": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "http.CreateGroupRequest": {
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "requireAdminApproval": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "http.JoinGroupRequest": {
                "properties": {
                    "inviteCode": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "http.ListGroupsResponse": {
                "properties": {
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/http.DetailsResponse"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "pagination": {
                        "$ref": "#/components/schemas/pagination.Meta"
                    }
                },
                "type": "object"
            },
...
            "response.ErrorResponse": {
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "Backend API service for Splittr bill splitting.",
        "title": "Splittr API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/groups": {
            "get": {
                "description": "Retrieve a cursor-paginated list of groups the current user belongs to.",
                "parameters": [
                    {
                        "description": "Items per page (max 100, default 20)",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Opaque cursor token from a previous response",
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ListGroupsResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List groups",
                "tags": [
                    "groups"
                ]
            },
            "post": {
                "description": "Create a new bill-splitting group with optional admin approval requirement.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/http.CreateGroupRequest",
                                        "summary": "request",
                                        "description": "Group creation data"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Group creation data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.Group"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create group",
                "tags": [
                    "groups"
                ]
            }
        },
        "/groups/join": {
            "post": {
                "description": "Join an existing group using its unique invite code.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "type": "object"
                                    },
                                    {
                                        "$ref": "#/components/schemas/http.JoinGroupRequest",
                                        "summary": "request",
                                        "description": "Group join data"
                                    }
                                ]
                            }
                        }
                    },
                    "description": "Group join data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.JoinResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Join group",
                "tags": [
                    "groups"
                ]
            }
        },
        "/groups/preview": {
            "get": {
                "description": "Get group details using invite code.",
                "parameters": [
                    {
                        "description": "Group invite code",
                        "in": "query",
                        "name": "inviteCode",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/domain.Preview"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get group preview",
                "tags": [
                    "groups"
                ]
            }
        },
...
        "/groups/{id}/feed": {
            "get": {
                "description": "Get a cursor-paginated timeline of events inside a group.",
                "parameters": [
                    {
                        "description": "Group ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Items per page (max 100, default 20)",
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Opaque cursor token from previous response",
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/activity.FeedResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Forbidden"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/response.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get group activity feed",
                "tags": [
                    "groups"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/v1"
        }
    ]
}